local testing & code re-structuring
